Created a library from TinyDB
Edited a few lines of code getLong was returning an int so i changed
that to long. When getting a double back there was a misspelling
theString was the name of the string and number was the name given in
Double.parse so I changed them both to number.

Otherwise exact same code just created as a library instead of just a
file.